Move grammar tests from e2e to unit
