+++ test/unit/grammar/cases.contractions.js
@@ -1,4 +1,4 @@
-module.exports.cases = {
+export const contractions = {
     Contractions: [
         {
             descr: "'s",
