+++ test/unit/index.html
@@ -5,7 +5,7 @@
     <title>Mocha Tests</title>
     <link href="https://unpkg.com/mocha@4.0.1/mocha.css" rel="stylesheet" />
     <script src="https://unpkg.com/mocha@4.0.1/mocha.js"></script>
-    <script src="https://unpkg.com/chai@4.1.2/chai.js"></script>;
+    <script src="https://unpkg.com/chai@4.1.2/chai.js"></script>
     <script src="../../lib/compromise.min.js"></script>
 </head>
 <body>
