+++ test/unit/index.js
@@ -4,3 +4,4 @@ import "./excluded-domains.test.js";
 import "./pronoun-replacement.test.js";
 import "./stopword-highlights.test.js";
 import "./word-replacement.test.js";
+import "./grammar/test-runner.js";
